CheckInherentsResult moved to primitives
